New idle mode, and some small changes
