+++ Assets/Resources/finishParticles.prefab
@@ -27,7 +27,7 @@ Transform:
   m_GameObject: {fileID: 8580816041480344387}
   serializedVersion: 2
   m_LocalRotation: {x: -0.7071068, y: 0, z: 0, w: 0.7071068}
-  m_LocalPosition: {x: 22.830189, y: 0, z: 0}
+  m_LocalPosition: {x: 1500.3279, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
