+++ Assets/Scenes/SampleScene.unity
@@ -2493,7 +2493,6 @@ MonoBehaviour:
   levelLength: {fileID: 104856853}
   delay: 0
   delayLimit: 0.25
-  timer: 0
   minWidth: 1
   maxWidth: 999
   expansionSpeed: 2
@@ -10034,8 +10033,6 @@ GameObject:
   - component: {fileID: 1724328861}
   - component: {fileID: 1724328860}
   - component: {fileID: 1724328859}
-  - component: {fileID: 1724328863}
-  - component: {fileID: 1724328864}
   - component: {fileID: 1724328865}
   m_Layer: 5
   m_Name: menu
@@ -10159,64 +10156,6 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1724328857}
   m_CullTransparentMesh: 1
---- !u!114 &1724328863
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1724328857}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d0b148fe25e99eb48b9724523833bab1, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  m_Delegates:
-  - eventID: 0
-    callback:
-      m_PersistentCalls:
-        m_Calls:
-        - m_Target: {fileID: 1724328864}
-          m_TargetAssemblyTypeName: smallButtonPanels, Assembly-CSharp
-          m_MethodName: ShowButton
-          m_Mode: 1
-          m_Arguments:
-            m_ObjectArgument: {fileID: 0}
-            m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-            m_IntArgument: 0
-            m_FloatArgument: 0
-            m_StringArgument: 
-            m_BoolArgument: 0
-          m_CallState: 2
-  - eventID: 1
-    callback:
-      m_PersistentCalls:
-        m_Calls:
-        - m_Target: {fileID: 1724328864}
-          m_TargetAssemblyTypeName: smallButtonPanels, Assembly-CSharp
-          m_MethodName: HideButton
-          m_Mode: 1
-          m_Arguments:
-            m_ObjectArgument: {fileID: 0}
-            m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-            m_IntArgument: 0
-            m_FloatArgument: 0
-            m_StringArgument: 
-            m_BoolArgument: 0
-          m_CallState: 2
---- !u!114 &1724328864
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1724328857}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0bb286d9cc306a34aa27a0f320d9ce3f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  panel: {fileID: 0}
 --- !u!114 &1724328865
 MonoBehaviour:
   m_ObjectHideFlags: 0
