+++ Assets/Scripts/JammerDash.Game/PlayerMovement.cs
@@ -12,6 +12,7 @@ using UnityEngine.InputSystem;
 using JammerDash.Tech;
 using JammerDash.Game.Player;
 using UnityEngine.InputSystem.Controls;
+using JammerDash.Difficulty;
 namespace JammerDash.Game.Player
 {
     public class PlayerMovement : MonoBehaviour 
@@ -408,7 +409,7 @@ namespace JammerDash.Game.Player
 
             // Calculate skill performance point
             float distanceInPercent = Mathf.Abs(cam.transform.position.x / FindObjectOfType<FinishLine>().transform.position.x);
-            GameScoreCalculator calc = new GameScoreCalculator(five, three, one, misses, combo, highestCombo, CustomLevelDataManager.Instance.diff, distanceInPercent);
+            ShinePerformance calc = new ShinePerformance(five, three, one, misses, combo, highestCombo, CustomLevelDataManager.Instance.diff, distanceInPercent);
             float skillPerformancePoint = calc.PerformanceScore;
             SPInt = Mathf.RoundToInt(skillPerformancePoint);
             if (Total > 0)
@@ -970,91 +971,5 @@ namespace JammerDash.Game.Player
         }
     }
 
-    public class GameScoreCalculator
-    {
-        private float _precision;
-        private float _sequenceEfficiency;
-        private float _versatility;
-        private float _completionSpeed;
-        private float _performanceScore;
-        public float _levelLength;
-        public int _perfectHits, _greatHits, _goodHits, _missedHits;
-        public int _currentCombo, _bestCombo;
-        private int _totalActions;
-        private float _gameDifficulty;
-        public GameScoreCalculator(int perfectHits, int greatHits, int goodHits, int missedHits, int currentCombo, int bestCombo, float gameDifficulty, float levelLength)
-        {
-            _perfectHits = perfectHits;
-            _greatHits = greatHits;
-            _goodHits = goodHits;
-            _missedHits = missedHits;
-            _currentCombo = currentCombo;
-            _bestCombo = bestCombo;
-            _totalActions = perfectHits + greatHits + goodHits + missedHits;
-            _gameDifficulty = gameDifficulty;
-            _levelLength = levelLength;
-
-            CalculatePrecision();
-            CalculateSequenceEfficiency();
-            CalculateVersatility(_perfectHits, _greatHits, _goodHits);
-            CalculateCompletionSpeed();
-            CalculatePerformanceScore();
-            CalculateLevelCompletionSpeed(levelLength);
-        }
-
-        private void CalculatePrecision()
-        {
-            _precision = Clamp((float)(_perfectHits + _greatHits + _goodHits) / _totalActions, 0, 1);
-        }
-
-        private void CalculateSequenceEfficiency()
-        {
-            _sequenceEfficiency = _bestCombo != 0 ? Clamp((float)_currentCombo / _bestCombo + 1, 0, 1) : 0;
-        }
-
-        private void CalculateVersatility(float high, float medium, float low)
-        {
-            
-            _versatility = high * 0.05f + medium * 0.025f + low * -0.1f + _missedHits * -0.5f;
-        }
-
-        private void CalculateCompletionSpeed()
-        {
-            _completionSpeed = Clamp01(CalculateLevelCompletionSpeed(_levelLength) / 1);
-        }
-
-        private float CalculateLevelCompletionSpeed(float length)
-        {
-
-            return 1.0f;
-        }
-
-        private void CalculatePerformanceScore()
-        {
-            float precisionWeight = 1.24f;
-
-            _performanceScore = Math.Max((
-                MathF.Pow(_precision, 2) * precisionWeight +
-                _sequenceEfficiency * 10 +
-                _versatility * 5 +
-                _completionSpeed +
-                _gameDifficulty * 0.35f) / 10,
-                0f
-            );
-        }
-
-        public float PerformanceScore => _performanceScore;
-
-        private static float Clamp(float value, float min, float max)
-        {
-            return Math.Max(min, Math.Min(max, value));
-        }
-
-        private static float Clamp01(float value)
-        {
-            return Clamp(value, 0, 1);
-        }
-
-
-    }
+   
 }
