+++ Assets/Scripts/JammerDash.Menus/KeybindUI.cs
@@ -2,56 +2,62 @@ using System;
 using UnityEngine;
 using UnityEngine.UI;
 using JammerDash;
-public class KeybindingUI : MonoBehaviour
+
+namespace JammerDash.Menus.Options
 {
-    public KeybindingManager keybindingManager;
-    public string keyName;
 
-    private void Start()
+    public class KeybindingUI : MonoBehaviour
     {
-        KeybindingManager.instance.LoadKeybindingsFromJson();
-        UpdateUI(this.GetComponent<Text>(), keyName);
-        GetComponentInParent<Button>().onClick.RemoveAllListeners();
-        GetComponentInParent<Button>().onClick.AddListener(Rebind);
-    }
+        public KeybindingManager keybindingManager;
+        public string keyName;
 
-    public void Rebind()
-    {
-        StartCoroutine(WaitForKeyPress(keyName));
-    }
-
-
-    private System.Collections.IEnumerator WaitForKeyPress(string actionName)
-    {
-        yield return new WaitUntil(() => Input.anyKeyDown);
-        foreach (KeyCode keycode in Enum.GetValues(typeof(KeyCode)))
+        private void Start()
         {
-            if (Input.GetKeyDown(keycode))
+            KeybindingManager.instance.LoadKeybindingsFromJson();
+            UpdateUI(this.GetComponent<Text>(), keyName);
+            GetComponentInParent<Button>().onClick.RemoveAllListeners();
+            GetComponentInParent<Button>().onClick.AddListener(Rebind);
+        }
+
+        public void Rebind()
+        {
+            StartCoroutine(WaitForKeyPress(keyName));
+        }
+
+
+        private System.Collections.IEnumerator WaitForKeyPress(string actionName)
+        {
+            yield return new WaitUntil(() => Input.anyKeyDown);
+            foreach (KeyCode keycode in Enum.GetValues(typeof(KeyCode)))
             {
-                KeybindingManager.RebindKey(actionName, keycode);
-                break;
+                if (Input.GetKeyDown(keycode))
+                {
+                    KeybindingManager.RebindKey(actionName, keycode);
+                    break;
+                }
+            }
+            UpdateUI(this.GetComponent<Text>(), keyName);
+        }
+        private void Update()
+        {
+            if (keybindingManager == null)
+            {
+                keybindingManager = KeybindingManager.instance;
+                KeybindingManager.instance.LoadKeybindingsFromJson();
             }
         }
-        UpdateUI(this.GetComponent<Text>(), keyName);
-    }
-    private void Update()
-    {
-        if (keybindingManager == null)
+
+        public void UpdateUI(Text text, string name)
         {
-            keybindingManager = KeybindingManager.instance;
-            KeybindingManager.instance.LoadKeybindingsFromJson();
+            text.text = KeybindingManager.GetBindingName(name);
+            Debug.Log("Updated UI with current bindings");
+        }
+
+        public void Save()
+        {
+
+            KeybindingManager.instance.SaveKeybindingsToJson();
         }
     }
 
-    public void UpdateUI(Text text, string name)
-    {
-        text.text = KeybindingManager.GetBindingName(name);
-        Debug.Log("Updated UI with current bindings");
-    }
-
-    public void Save()
-    {
-
-        KeybindingManager.instance.SaveKeybindingsToJson();
-    }
 }
