+++ Assets/Scripts/JammerDash/Difficulty.cs
@@ -1,13 +1,14 @@
+using System;
 using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace JammerDash
+namespace JammerDash.Difficulty
 {
 
-    public class Difficulty : MonoBehaviour
+    public class Calculator : MonoBehaviour
     {
         public static float CalculateDifficulty(List<GameObject> cubes, List<GameObject> saws, List<GameObject> longCubes, Slider hp, Slider size, int[] cubeCountsPerY, Vector2[] cubePositions, float clickTimingWindow)
         {
@@ -154,6 +155,83 @@ namespace JammerDash
             }
 
             return objects;
+        }
+
+
+    }
+    public class ShinePerformance
+    {
+        private float _precision;
+        private float _sequenceEfficiency;
+        private float _versatility;
+        private float _performanceScore;
+        public float _levelLength;
+        public int _perfectHits, _greatHits, _goodHits, _missedHits;
+        public int _currentCombo, _bestCombo;
+        private int _totalActions;
+        private float _gameDifficulty;
+        public ShinePerformance(int perfectHits, int greatHits, int goodHits, int missedHits, int currentCombo, int bestCombo, float gameDifficulty, float levelLength)
+        {
+            _perfectHits = perfectHits;
+            _greatHits = greatHits;
+            _goodHits = goodHits;
+            _missedHits = missedHits;
+            _currentCombo = currentCombo;
+            _bestCombo = bestCombo;
+            _totalActions = perfectHits + greatHits + goodHits + missedHits;
+            _gameDifficulty = gameDifficulty;
+            _levelLength = levelLength;
+
+            CalculatePrecision();
+            CalculateSequenceEfficiency();
+            CalculateVersatility(_perfectHits, _greatHits, _goodHits);
+            CalculatePerformanceScore();
+        }
+
+        private void CalculatePrecision()
+        {
+            _precision = Clamp((float)(_perfectHits + _greatHits + _goodHits) / _totalActions, 0, 1);
+        }
+
+        private void CalculateSequenceEfficiency()
+        {
+            _sequenceEfficiency = _bestCombo != 0 ? Clamp((float)_currentCombo / _bestCombo + 1, 0, 1) : 0;
+        }
+
+        private void CalculateVersatility(float high, float medium, float low)
+        {
+
+            _versatility = high * 0.05f + medium * 0.025f + low * -0.1f + _missedHits * -0.5f;
+        }
+
+      
+
+      
+
+        private void CalculatePerformanceScore()
+        {
+            float precisionWeight = 1.24f;
+
+            _performanceScore = Math.Max((
+                MathF.Pow(_precision, 2) * precisionWeight +
+                _sequenceEfficiency * 10 +
+                _versatility * 5 +
+                _levelLength / 100 +
+                _gameDifficulty * 0.35f) / 10,
+                0f
+            );
+        }
+
+        public float PerformanceScore => _performanceScore;
+
+        private static float Clamp(float value, float min, float max)
+        {
+            return Math.Max(min, Math.Min(max, value));
+        }
+
+        private static float Clamp01(float value)
+        {
+            return Clamp(value, 0, 1);
         }
