+++ Assets/StreamingAssets/tips.txt
@@ -11,7 +11,7 @@ The editor guide has a lot of useful things. Make sure to check it out!
 Shift + left click changes the long cube width in the editor.
 To get a perfect score, destroy each box with as little mistakes as possible.
 Switch between fullscreen and windowed with "Alt+Enter"!
-Take occassional breaks. Jammer Dash isn't going to change your life drastically.
+Take occassional breaks. Jammer Dash isn't going to go away.
 Grinding to be the best isn't the goal. The goal is to have fun!
 Check out the latest news in the Discord!
 Are you a fan of snow? Enable the snow option in the options to have snow falling in the menu!!
